day03/src/requirements.txt
ansible==9.4.0
ansible-core==2.16.6
beautifulsoup4==4.12.3
cffi==1.16.0
cryptography==42.0.5
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.22
PyYAML==6.0.1
redis==5.0.3
resolvelib==1.0.1
soupsieve==2.5
